refactor: use proper URLs for routes
