+++ src/app/login/applicator-lead.tsx
@@ -1,20 +1,36 @@
 import React, { useEffect, useState } from 'react'
-import { View, Text, Alert, ScrollView } from 'react-native'
+import {
+  View,
+  Text,
+  Alert,
+  ScrollView,
+  TextInput,
+  Pressable,
+  ActivityIndicator,
+} from 'react-native'
 import * as LocalAuthentication from 'expo-local-authentication'
 import NetInfo from '@react-native-community/netinfo'
 import { router } from 'expo-router'
 import { z } from 'zod'
-import { useForm } from 'react-hook-form'
+import { Controller, useForm } from 'react-hook-form'
 import { useUser } from '@/contexts/UserContext'
 import { zodResolver } from '@hookform/resolvers/zod'
-import FormControl from './FormControl'
 import { doLogin } from '@/services/onlineServices/authenticate'
 import { Snackbar } from 'react-native-paper'
 import AsyncStorage from '@react-native-async-storage/async-storage'
 import { useDevice } from '@/features/device'
 import { useApplicator } from '@/contexts/ApplicatorContext'
+import { FontAwesome } from '@expo/vector-icons'
 
 const authSchema = z.object({
+  email: z
+    .string({
+      required_error: 'Email é obrigatório',
+      invalid_type_error: 'Email precisa ser uma string',
+    })
+    .min(2, 'Email precisa ter pelo menos 2 caracteres')
+    .email('Email precisa ser um email válido'),
+
   password: z
     .string({
       required_error: 'Senha é obrigatório',
@@ -25,14 +41,14 @@ const authSchema = z.object({
 
 export type AuthFormData = z.infer<typeof authSchema>
 
-const Login = () => {
-  const { device, refetchDevice } = useDevice()
-  const { fetchApplicatorData } = useApplicator()
-  const { loginUser } = useUser()
+const ApplicatorLeadLogin = () => {
   const [showPassword, setShowPassword] = useState(false)
   const [buttonLoading, setButtonLoading] = useState(false)
   const [visibleOK, setVisibleOK] = useState(false)
   const [visibleERROR, setVisibleERROR] = useState(false)
+
+  const { refetchDevice } = useDevice()
+  const { fetchApplicatorData } = useApplicator()
 
   const {
     control,
@@ -42,23 +58,28 @@ const Login = () => {
     resolver: zodResolver(authSchema),
   })
 
+  const { loginUser } = useUser()
+
   const onDismissSnackBarOK = () => setVisibleOK(false)
+
   const onDismissSnackBarERROR = () => setVisibleERROR(false)
 
-  const onSubmit = handleSubmit(async (data: AuthFormData) => {
+  const onSubmit = handleSubmit(async (data) => {
+    // FIXME: why do we need these refetches?
     await refetchDevice()
     await fetchApplicatorData()
     try {
       setButtonLoading(true)
 
-      const response = await doLogin(device.factory_id, data.password)
+      const response = await doLogin(data.email, data.password)
 
-      if (response?.user) {
+      if (response && response.user.is_staff) {
         setVisibleOK(!visibleOK)
-        AsyncStorage.setItem('token_service_id', data.password)
+
         setTimeout(() => {
           setVisibleOK(!visibleOK)
-          router.navigate('/login/applicatorCpfVerificate')
+          router.navigate('/login/applicator-cpf-verify')
+          AsyncStorage.removeItem('token_service_id')
         }, 1000)
 
         loginUser(response.user)
@@ -114,21 +135,82 @@ const Login = () => {
       <View className="container h-screen flex-1 bg-white">
         <View className="flex-1 justify-center gap-5">
           <Text className="text-4xl font-bold text-black">
-            Entrar com o verificador de serviço
+            Entrar como líder de equipe
           </Text>
           <Text className="text-xl text-black">
             Insira seus dados e explore todas as funcionalidades que temos a
             oferecer
           </Text>
+          <View className="gap-5">
+            <Controller
+              control={control}
+              render={({ field: { onChange, onBlur, value } }) => (
+                <View className="relative">
+                  <TextInput
+                    onBlur={onBlur}
+                    onChangeText={(value) => onChange(value)}
+                    value={value}
+                    placeholder="Email"
+                    className=" rounded-md border border-zinc-700/20 p-2 pl-4"
+                  />
+                  {errors && (
+                    <Text className="absolute -bottom-5 text-sm text-red-500">
+                      {errors.email?.message}
+                    </Text>
+                  )}
+                </View>
+              )}
+              name="email"
+              rules={{ required: true }}
+              defaultValue=""
+            />
+            <Controller
+              control={control}
+              render={({ field: { onChange, onBlur, value } }) => (
+                <View className="relative">
+                  <TextInput
+                    onBlur={onBlur}
+                    onChangeText={(value) => onChange(value)}
+                    value={value}
+                    placeholder="Senha"
+                    secureTextEntry={!showPassword}
+                    className=" rounded-md border border-zinc-700/20 p-2 pl-4"
+                  />
+                  <Pressable className=" absolute right-5 top-3">
+                    <FontAwesome
+                      name={showPassword ? 'eye-slash' : 'eye'}
+                      size={24}
+                      color="black"
+                      onPress={() => setShowPassword(!showPassword)}
+                    />
+                  </Pressable>
+                  {errors && (
+                    <Text className="absolute -bottom-5 text-sm text-red-500">
+                      {errors.password?.message}
+                    </Text>
+                  )}
+                </View>
+              )}
+              name="password"
+              rules={{ required: true }}
+              defaultValue=""
+            />
 
-          <FormControl
-            buttonLoading={buttonLoading}
-            control={control}
-            errors={errors}
-            onSubmit={onSubmit}
-            setShowPassword={setShowPassword}
-            showPassword={showPassword}
-          />
+            {buttonLoading ? (
+              <Pressable className="rounded-md bg-zinc-500 p-3">
+                <ActivityIndicator size="small" color="#fff" />
+              </Pressable>
+            ) : (
+              <Pressable
+                onPress={onSubmit}
+                className="rounded-md bg-zinc-500 p-3"
+              >
+                <Text className="text-md text-center font-bold text-white">
+                  ENTRAR
+                </Text>
+              </Pressable>
+            )}
+          </View>
         </View>
       </View>
       <Snackbar
@@ -143,7 +225,7 @@ const Login = () => {
         }}
       >
         <Text className="text-zinc-700">
-          Líder de equipe logado com sucesso.
+          Aplicador lider logado com sucesso.
         </Text>
       </Snackbar>
       <Snackbar
@@ -165,4 +247,4 @@ const Login = () => {
   )
 }
 
-export default Login
+export default ApplicatorLeadLogin
